Distributed-LMS-with-LLM-based-Tutoring-main/AOS-LMS-M3/lms-client/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.28.1
setuptools==74.1.2
